refactor process cleanup helpers
